requirements.txt
requests
tweepy
pyyaml
pyspark
pandas
textblob